Refactor: Move model-generating files to the model_gen directory
