+++ model_gen/naive.py
